extract userProvider, add documentProvider (wip)
